Restructuring, coroutine supervisor
Went over the code with a sponge and soapy water and cleaned some of it up.
Also, since the /place experiment concluded, I can't test this script anymore,
so this is the final version.

Unfortunately the script wasn't usable until around 12-24 hours after the
experiment had started, and even then I accidentally deleted some of the
historical images, so the archive I collected using this script is quite poor.